requirements.txt
Django>=6.0.0
Pillow>=10.0.0
django-environ>=0.11.0
django-tailwind>=3.7.0
whitenoise>=6.5.0
psycopg2-binary>=2.9.9
djangorestframework>=3.15.0
django-filter>=24.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-allauth>=65.0.0
meilisearch==0.40.0
